+++ Study.Lab2.Logic.UnitTests/kinkiss1/ServerRequestServiceTests.cs
@@ -1,6 +1,6 @@
 using Moq;
 using Study.Lab2.Logic.Interfaces.kinkiss1;
-using Study.Lab2.Logic.kinkiss1.DtoModels;
+using Study.Lab2.Logic.UnitTests.kinkiss1;
 
 namespace Study.Lab2.Logic.kinkiss1;
