+++ Study.Lab2.Logic/kinkiss1/DtoModels/CatFactResponceDto.cs
@@ -1,13 +1,12 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace Study.Lab2.Logic.kinkiss1.DtoModels
-{
-    public class CatFactResponseDto
-    {
-        [JsonPropertyName("fact")]
-        public string Fact { get; init; }
+namespace Study.Lab2.Logic.kinkiss1.DtoModels;
 
-        [JsonPropertyName("length")]
-        public int Length { get; init; }
-    }
+public class CatFactResponseDto
+{
+    [JsonPropertyName("fact")]
+    public string Fact { get; init; }
+
+    [JsonPropertyName("length")]
+    public int Length { get; init; }
 }
