+++ Study.Lab2.Logic/kinkiss1/DtoModels/CatFactTranslatedResponseDto.cs
@@ -1,16 +1,14 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace Study.Lab2.Logic.kinkiss1.DtoModels
+namespace Study.Lab2.Logic.kinkiss1.DtoModels;
+public class CatFactTranslatedResponseDto
 {
-    public class CatFactTranslatedResponseDto
-    {
-        [JsonPropertyName("fact")]
-        public string Fact { get; init; }
+    [JsonPropertyName("fact")]
+    public string Fact { get; init; }
 
-        [JsonPropertyName("length")]
-        public int Length { get; init; }
+    [JsonPropertyName("length")]
+    public int Length { get; init; }
 
-        [JsonPropertyName("перевод")]
-        public string Translate { get; init; }
-    }
+    [JsonPropertyName("перевод")]
+    public string Translate { get; init; }
 }
