+++ Study.Lab2.Logic/kinkiss1/DtoModels/KanyeRestResponseDto.cs
@@ -1,10 +1,8 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace Study.Lab2.Logic.kinkiss1.DtoModels
+namespace Study.Lab2.Logic.kinkiss1.DtoModels;
+public class KanyeRestResponseDto
 {
-    public class KanyeRestResponseDto
-    {
-        [JsonPropertyName("quote")]
-        public string Quote { get; init; }
-    }
+    [JsonPropertyName("quote")]
+    public string Quote { get; init; }
 }
